Set minimal radius of point parameters (when specified as percent)

+++ src/KeypointList.cpp
@@ -49,7 +49,7 @@ void KeypointList::clear()
 QPointF KeypointList::position(int n) const
 {
   const Keypoint & kp = _keypoints[n];
-  return {kp.x, kp.y};
+  return {static_cast<qreal>(kp.x), static_cast<qreal>(kp.y)};
 }
 
 QColor KeypointList::color(int n) const
