+++ src/Widgets/PreviewWidget.cpp
@@ -251,7 +251,7 @@ void PreviewWidget::paintKeypoints(QPainter & painter)
 
   QRect visibleRect = rect() & _imagePosition;
   KeypointList::reverse_iterator it = _keypoints.rbegin();
-  int index = _keypoints.size() - 1;
+  int index = static_cast<int>(_keypoints.size() - 1);
   while (it != _keypoints.rend()) {
     if (!it->isNaN()) {
       const KeypointList::Keypoint & kp = *it;
